Refactor catalog-related data out of evaluation context.
Moving catalog-related evaluation data (class, defined type, and node
definitions) into the catalog itself.

This makes the evaluation context more generic.
