+++ lib/include/puppet/runtime/definition_scanner.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include "../ast/syntax_tree.hpp"
-#include "context.hpp"
+#include "catalog.hpp"
 #include <memory>
 
 namespace puppet { namespace runtime {
@@ -13,15 +13,15 @@ namespace puppet { namespace runtime {
     /**
      * Represents the runtime definition scanner.
      * This type is responsible for scanning a syntax tree and
-     * adding classes, defined types, and nodes to the evaluation context.
+     * defining classes, defined types, and nodes in a catalog.
      */
     struct definition_scanner
     {
         /**
          * Constructs a definition scanner.
-         * @param context The evaluation context.
+         * @param catalog The catalog to populate with definitions.
          */
-        explicit definition_scanner(runtime::context& context);
+        explicit definition_scanner(runtime::catalog& catalog);
 
         /**
          * Scans the given compilation context's syntax tree for definition.
@@ -30,7 +30,7 @@ namespace puppet { namespace runtime {
         void scan(std::shared_ptr<compiler::context> compilation_context);
 
      private:
-        runtime::context& _context;
+        runtime::catalog& _catalog;
     };
 
 }}  // puppet::runtime
